fix: top level await (chrome 129)
