+++ package.json
@@ -37,7 +37,7 @@
     "@iconify-emoji/twemoji": "^1.0.2",
     "@iconify/json": "^2.2.170",
     "@iconify/utils": "^2.1.22",
-    "@nuxt/devtools": "^1.4.2",
+    "@nuxt/devtools": "^1.5.0",
     "@nuxt/test-utils": "^3.14.2",
     "@nuxtjs/color-mode": "^3.4.4",
     "@nuxtjs/i18n": "^8.5.3",
